refactor: move blocking code to new file
