+++ wal/src/blocking.rs
@@ -1,0 +1,2 @@
+mod reader;
+pub use reader::{Error as ReaderError, Result as ReaderResult, SegmentFileReader};
